time::date: rename /unit/step/, /unitdiff()/hop()/
This is an effort to improve naming. Simple words or short phrases which
serve as a metaphor are preferred.

"unit" did not precisely evoke its intended meaning. It did not
encapsulate the variability of months, years, etc. "step" proved better.

"unitdiff" hence required renaming. Usage of "diff" is discouraged since
we want a convention where "diff" deals with timescalar (duration) code.
Instead of "stepdiff", "hop" was chosen. One "hops" by a single step
from A to B, and the number of traversed steps is returned.

Breaking-change: time::date
Signed-off-by: Byron Torres <b@torresjrjr.com>