cmd/hare: rebuild on binary change based on mtime
previously, the complete binary was hashed. this takes a very long time at
least on openbsd, mostly because the size of $(CC) is 96.6M, which resulted
in a time increase for a simple "hare run" from 0.1 seconds to 3 seconds (fast
machine) or 15 seconds (slow machine).

using the mtime of the binary has almost the same effect but is much faster.

Signed-off-by: Lorenz (xha) <me@xha.li>