time::date: repurpose tz() to tzdb()
The function now only accepts a timezone identifier, simplifying the
interface and lessening the chance of user error.

The name "tzdb" better reflects the function's purpose, and emphasizes
the presence of a database. "tz" implies a generic timezone loader;
such a function is better implemented by hand by the user.

Signed-off-by: Byron Torres <b@torresjrjr.com>