unix::hosts: handle nomem
net::dial::error is updated to include nomem. It's added here rather
than added to the signature of various dial functions because it tends
to be bubbled up through abstractions via service handlers (i.e.
dial_tcp) in a manner similar to io::error, and generally indicates a
failure deeper in the stack (in this case it would be an OOM condition
while parsing /etc/hosts).

Breaking-change: unix::hosts::
Signed-off-by: Drew DeVault <sir@cmpwn.com>